feat: Implemented full encryption and decryption with padding
This is an implementation which fixes the issue with padding. The
padding now adds at least 16 bytes of padded data, and the padding
will always be at the end of the file. The padding is removed when
decrypting.

Added tests for full encryption and decryption with padding for
both text files and binary files.

Fixed minor improvvement when it came to slicing for improved
readability of the code.

Breaking changes: Old encrypted data won't work with this version

Resolves: None